==> 1-Local/2-Web_Scraping.ipynb <==
# **Obtención de registro de recorridos de taxis**

#### Para obtener los datos de recorridos de taxis amarillos y verdes en la Ciudad de New York, realizaremos Web Scraping a la [página](http://www.nyc.gov/html/tlc/html/about/trip_record_data.shtml) provista en el [diccionario de datos](Datasets/data_dictionary_trip_records_yellow.pdf).

### Importamos las librerías necesarias.

import requests
from bs4 import BeautifulSoup
from Funciones import *
import os
import re
import pandas as pd

### Efectuamos Web Scrapping para obtener los enlaces de registros de viajes en taxis.

# Guardamos la URL en una variable
url = 'https://www.nyc.gov/site/tlc/about/tlc-trip-record-data.page'

# Realizamos la solicitud HTTP
response = requests.get(url)

# Verificamos si la solicitud fue exitosa
if response.status_code == 200:
    # Analizamos el contenido HTML con BeautifulSoup
    soup = BeautifulSoup(response.text, 'html.parser')

    # Buscamos el elemento <div> con la clase "faq-answers" y el id "faq2023"
    target_div = soup.find('div', {'class': 'faq-answers', 'id': 'faq2023'})

    if target_div:
        # Buscamos los enlaces <a> dentro de target_div que tienen title="Yellow Taxi Trip Records" o "Green Taxi Trip Records"
        taxi_links = target_div.find_all('a', title=['Yellow Taxi Trip Records', 'Green Taxi Trip Records'])

        if taxi_links:
            print("Enlaces con el título 'Yellow Taxi Trip Records' o 'Green Taxi Trip Records':")
            links = []
            for link in taxi_links:
                links.append(link.get('href'))
                print(link.get('href'))  # Imprime el enlace URL
        else:
            print("No se encontraron enlaces con los títulos 'Yellow Taxi Trip Records' o 'Green Taxi Trip Records' dentro de 'faq-answers'.")

    else:
        print("No se encontró el elemento con la clase 'faq-answers' y el id 'faq2023'.")

else:
    print("La solicitud no fue exitosa. Código de estado:", response.status_code)

### Habiendo visualizado la estructura de los enlaces, procedemos a crear una carpeta donde se almacenarán los archivos que descargaremos.

# Creamos una arpeta de destino para guardar los archivos descargados
carpeta_destino = "Datasets/2-yellow_and_green_taxis/"

# Verificamos si la carpeta de destino existe, si no existe la creamos
if not os.path.exists(carpeta_destino):
    os.makedirs(carpeta_destino)

# Obtenemos la lista de archivos en la carpeta de destino
archivos_locales = os.listdir(carpeta_destino)

# Generamos dinámicamente los enlaces para los meses del 01 al 10
meses = ['01', '02', '03', '04', '05', '06', '07', '08', '09', '10']

# Guardamos la URL en una variable
url = "https://www.nyc.gov/site/tlc/about/tlc-trip-record-data.page"

# Realizamos la solicitud HTTP a la página
response = requests.get(url)

# Verificamos si la solicitud fue exitosa (código de estado 200)
if response.status_code == 200:
    # Parsear el contenido HTML de la página
    soup = BeautifulSoup(response.text, 'html.parser')

    # Encontramos todos los enlaces que contienen 'yellow' o 'green' y tienen el formato esperado para 2023
    links_pagina = {a['href'] for a in soup.find_all('a', href=True) if re.match(r'^.*(yellow|green)_tripdata_2023-\d{2}\.parquet$', a['href'])}

    # Combinamos enlaces generados dinámicamente con los obtenidos de la página
    links_totales = links_pagina.copy()  # Creamos una copia para mantener los enlaces originales
    for mes in meses:
        enlace_yellow = f"https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_2023-{mes}.parquet"
        enlace_green = f"https://d37ci6vzurychx.cloudfront.net/trip-data/green_tripdata_2023-{mes}.parquet"
        links_totales.add(enlace_yellow)
        links_totales.add(enlace_green)

    # Realizamos la descarga de los archivos
    for link in sorted(links_totales):
        nombre_archivo = link.split("/")[-1]  # Obtenemos el nombre del archivo desde la URL
        archivo_destino = os.path.join(carpeta_destino, nombre_archivo)

        if nombre_archivo not in archivos_locales:
            response = requests.get(link)

            if response.status_code == 200:
                with open(archivo_destino, 'wb') as archivo:
                    archivo.write(response.content)
                print(f"Archivo descargado: {nombre_archivo}")
            else:
                print(f"No se pudo descargar el archivo: {nombre_archivo}. Código de estado: {response.status_code}")
        else:
            print(f"El archivo {nombre_archivo} ya existe localmente.")
else:
    print(f"No se pudo acceder a la página. Código de estado: {response.status_code}")


### _**Unimos los dataframes mensuales de taxis verdes en un solo dataframe.**_

enero_g = pd.read_parquet("Datasets/2-yellow_and_green_taxis/green_tripdata_2023-01.parquet")
febrero_g = pd.read_parquet("Datasets/2-yellow_and_green_taxis/green_tripdata_2023-02.parquet")
marzo_g = pd.read_parquet("Datasets/2-yellow_and_green_taxis/green_tripdata_2023-03.parquet")
abril_g = pd.read_parquet("Datasets/2-yellow_and_green_taxis/green_tripdata_2023-04.parquet")
mayo_g = pd.read_parquet("Datasets/2-yellow_and_green_taxis/green_tripdata_2023-05.parquet")
junio_g = pd.read_parquet("Datasets/2-yellow_and_green_taxis/green_tripdata_2023-06.parquet")
julio_g = pd.read_parquet("Datasets/2-yellow_and_green_taxis/green_tripdata_2023-07.parquet")
agosto_g = pd.read_parquet("Datasets/2-yellow_and_green_taxis/green_tripdata_2023-08.parquet")
septiembre_g = pd.read_parquet("Datasets/2-yellow_and_green_taxis/green_tripdata_2023-09.parquet")
octubre_g = pd.read_parquet("Datasets/2-yellow_and_green_taxis/green_tripdata_2023-10.parquet")

# Listamos los DataFrames
dfs = [enero_g, febrero_g, marzo_g, abril_g, mayo_g, junio_g, julio_g, agosto_g, septiembre_g, octubre_g]

# Concateamos para crear uno solo
green_taxis = pd.concat(dfs, ignore_index=True)
green_taxis.info()

# Guardamos columnas a eliminar en una variable
columnas_eliminar = ['store_and_fwd_flag', 'fare_amount', 'extra', 'mta_tax', 'tolls_amount', 'improvement_surcharge',
                    'ehail_fee', 'congestion_surcharge', 'trip_type', 'VendorID', 'RatecodeID', 'payment_type']

# Eliminamos las columnas de la lista
green_taxis.drop(columns=columnas_eliminar, inplace=True)

# Renombramos las columnas
green_taxis.rename(columns={'lpep_pickup_datetime': 'pickup_datetime','lpep_dropoff_datetime': 'dropoff_datetime',
                            'total_amount': 'amount'}, inplace=True)
green_taxis.info()

green_taxis.head()

### En New York la propina (tip_amount) se considera una costumbre. Conforme a lo detallado en el diccionario provisto, el pago total incluye impuestos pero no propina asi que procedemos a sumar las columnas 'amount' y 'tip_amount'.

# Sumamos las columnas amount y tip_amount para crear la columna total_amount
green_taxis['total_amount'] = green_taxis['amount'] + green_taxis['tip_amount']

# Mostrar las primeras filas del DataFrame resultante
green_taxis.head()

# Eliminamos las columnas amount y tip_amount al no ser necesarias
green_taxis.drop(columns=['amount', 'tip_amount'], inplace=True)
green_taxis.info()

# Revisamos el contenido de la columna passenger_count
green_taxis['passenger_count'].value_counts()

# Verificamos nulos
green_taxis['passenger_count'].isnull().sum()

# Verificamos la composicion del DataFrame cuando passenger_count es nulo para no tomar decisiones erroneas
verificar = green_taxis[green_taxis['passenger_count'].isnull()]
verificar.head()

### Teniendo en cuenta que hay ocasiones en que no se informa de cantidad de pasajeros pero si hay un registro de pago, procedemos a completar los valores nulos de la columna 'passenger_count' con el valor 1.

# Completamos los valores faltantes de la columna passenger_count con el valor 1
green_taxis['passenger_count'].fillna(1, inplace=True)
green_taxis['passenger_count'].isnull().sum()

# Eliminamos las filas donde passenger_count es igual a 0.0
green_taxis = green_taxis[green_taxis['passenger_count'] != 0.0]
green_taxis['passenger_count'].value_counts()

# Revisamos el contenido de la columna trip_distance
green_taxis['trip_distance'].value_counts()

# Eliminamos las filas donde trip_distance es igual a 0.00
green_taxis = green_taxis[green_taxis['trip_distance'] != 0.00]
green_taxis['trip_distance'].value_counts()

# Creamos una columna llamada service_type con el valor 'green'
green_taxis['service_type'] = 'green'
green_taxis.head(2)

### _**Unimos los dataframes mensuales de taxis amarillos en un solo dataframe.**_

enero_y = pd.read_parquet("Datasets/2-yellow_and_green_taxis/yellow_tripdata_2023-01.parquet")
febrero_y = pd.read_parquet("Datasets/2-yellow_and_green_taxis/yellow_tripdata_2023-02.parquet")
marzo_y = pd.read_parquet("Datasets/2-yellow_and_green_taxis/yellow_tripdata_2023-03.parquet")
abril_y = pd.read_parquet("Datasets/2-yellow_and_green_taxis/yellow_tripdata_2023-04.parquet")
mayo_y = pd.read_parquet("Datasets/2-yellow_and_green_taxis/yellow_tripdata_2023-05.parquet")
junio_y = pd.read_parquet("Datasets/2-yellow_and_green_taxis/yellow_tripdata_2023-06.parquet")
julio_y = pd.read_parquet("Datasets/2-yellow_and_green_taxis/yellow_tripdata_2023-07.parquet")
agosto_y = pd.read_parquet("Datasets/2-yellow_and_green_taxis/yellow_tripdata_2023-08.parquet")
septiembre_y = pd.read_parquet("Datasets/2-yellow_and_green_taxis/yellow_tripdata_2023-09.parquet")
octubre_y = pd.read_parquet("Datasets/2-yellow_and_green_taxis/yellow_tripdata_2023-10.parquet")

# Listamos los DataFrames
dfs = [enero_y, febrero_y, marzo_y, abril_y, mayo_y, junio_y, julio_y, agosto_y, septiembre_y, octubre_y]

# Concatenamos para crear uno solo
yellow_taxis = pd.concat(dfs, ignore_index=True)
yellow_taxis.info()

# Guardamos columnas a eliminar en una variable
columnas_eliminar = ['store_and_fwd_flag', 'fare_amount', 'extra', 'mta_tax', 'tolls_amount', 'improvement_surcharge',
                    'congestion_surcharge', 'airport_fee', 'Airport_fee', 'VendorID', 'RatecodeID', 'payment_type']

# Eliminamos las columnas de la lista
yellow_taxis.drop(columns=columnas_eliminar, inplace=True)

# Renombramos las columnas
yellow_taxis.rename(columns={'tpep_pickup_datetime': 'pickup_datetime','tpep_dropoff_datetime': 'dropoff_datetime',
                             'total_amount': 'amount'}, inplace=True)
yellow_taxis.info()

yellow_taxis.head()

### En New York la propina (tip_amount) se considera una costumbre. Conforme a lo detallado en el diccionario provisto, el pago total incluye impuestos pero no propina asi que procedemos a sumar las columnas 'amount' y 'tip_amount'.

# Sumamos las columnas amount y tip_amount para crear la columna total_amount
yellow_taxis['total_amount'] = yellow_taxis['amount'] + yellow_taxis['tip_amount']

# Mostrar las primeras filas del DataFrame resultante
yellow_taxis.head()

# Eliminamos las columnas amount y tip_amount al no ser necesarias
yellow_taxis.drop(columns=['amount', 'tip_amount'], inplace=True)
yellow_taxis.info()

# Revisamos los valores de la columna passenger_count
yellow_taxis['passenger_count'].value_counts()

# Verificamos nulos
yellow_taxis['passenger_count'].isnull().sum()

# Verificamos la composición del DataFrame cuando passenger_count es nulo para no tomar decisiones erroneas
verificar2 = yellow_taxis[yellow_taxis['passenger_count'].isnull()]
verificar2.head()

### Teniendo en cuenta que hay ocasiones en que no se informa de cantidad de pasajeros pero si hay un registro de pago, procedemos a completar los valores nulos de la columna 'passenger_count' con el valor 1.

# Completamos los valores faltantes de la columna passenger_count con el valor 1
yellow_taxis['passenger_count'].fillna(1, inplace=True)
yellow_taxis['passenger_count'].isnull().sum()

# Eliminamos las filas donde passenger_count es igual a 0.0
yellow_taxis = yellow_taxis[yellow_taxis['passenger_count'] != 0.0]
yellow_taxis['passenger_count'].value_counts()

# Revisamos los valores de la columna trip_distance
yellow_taxis['trip_distance'].value_counts()

# Eliminamos las filas donde trip_distance es igual a 0.00
yellow_taxis = yellow_taxis[yellow_taxis['trip_distance'] != 0.00]
yellow_taxis['trip_distance'].value_counts()

# Creamos una columna llamada service_type con el valor 'yellow'
yellow_taxis['service_type'] = 'yellow'
yellow_taxis.head(2)

### Teniendo los dataframes de taxis amarillos y verdes y habiendo creado una columna que los distinga, procedemos a unirlos en un solo dataframe.

# Concatenamos los DataFrames de yellow y green taxis
taxis = pd.concat([yellow_taxis, green_taxis], ignore_index=True)
taxis.info()

# Aplicamos la función para crear nuevas columnas
taxis['pickup_borough'] = taxis['PULocationID'].apply(get_borough)
taxis['dropoff_borough'] = taxis['DOLocationID'].apply(get_borough)

taxis.pickup_borough.value_counts()

taxis.dropoff_borough.value_counts()

# Eliminamos todas las filas donde pickup_borough o dropoff_borough sea igual a 'Unknown'
taxis = taxis[(taxis['pickup_borough'] != 'Unknown') & (taxis['dropoff_borough'] != 'Unknown')]
taxis.info()

# Cambiamos el tipo de dato de 'passenger_count' a int
taxis['passenger_count'] = taxis['passenger_count'].astype(int)
taxis.info()

### Guardamos el dataframe de taxis en un archivo parquet.

taxis.to_parquet("Datasets/3-Normalizados/taxis_2023.parquet")

import pandas as pd
import boto3
from io import BytesIO
from botocore.exceptions import NoCredentialsError

def read_parquet_from_s3(bucket_name, key):
    s3 = boto3.client('s3')
    try:
        response = s3.get_object(Bucket=bucket_name, Key=key)
        parquet_data = response['Body'].read()
        df = pd.read_parquet(BytesIO(parquet_data))
        return df
    except NoCredentialsError:
        print('Credentials not available')

def etl_taxis(df, service_type):
    # Guardamos columnas a eliminar en una variable
    columnas_eliminar = ['store_and_fwd_flag', 'fare_amount', 'extra', 'mta_tax', 'tolls_amount', 'improvement_surcharge',
                        'ehail_fee', 'congestion_surcharge', 'trip_type', 'VendorID', 'RatecodeID', 'payment_type']

    # Eliminamos las columnas de la lista
    df.drop(columns=columnas_eliminar, inplace=True)

    # Renombramos las columnas
    df.rename(columns={'lpep_pickup_datetime': 'pickup_datetime', 'lpep_dropoff_datetime': 'dropoff_datetime',
                       'tpep_pickup_datetime': 'pickup_datetime', 'tpep_dropoff_datetime': 'dropoff_datetime',
                       'total_amount': 'amount'}, inplace=True)

    # Sumamos las columnas amount y tip_amount para crear la columna total_amount
    df['total_amount'] = df['amount'] + df['tip_amount']

    # Eliminamos las columnas amount y tip_amount al no ser necesarias
    df.drop(columns=['amount', 'tip_amount'], inplace=True)

    # Completamos los valores faltantes de la columna passenger_count con el valor 1
    df['passenger_count'].fillna(1, inplace=True)

    # Eliminamos las filas donde passenger_count es igual a 0.0
    df = df[df['passenger_count'] != 0.0]

    # Eliminamos las filas donde trip_distance es igual a 0.00
    df = df[df['trip_distance'] != 0.00]

    # Creamos una columna llamada service_type con el valor proporcionado
    df['service_type'] = service_type

    return df

def get_borough(location_id):
    # ... (código de la función get_borough)
    return 'Unknown'

def lambda_handler(event, context):
    bucket_name = "web-scrap-bucket"
    service_types = ['yellow', 'green']
    dfs = []

    for service_type in service_types:
        service_dfs = []
        for month in range(1, 13):
            file_key = f"{service_type}_tripdata_2023-{str(month).zfill(2)}.parquet"
            df = read_parquet_from_s3(bucket_name, file_key)
            df_transformed = etl_taxis(df, service_type)
            service_dfs.append(df_transformed)

        taxis_data = pd.concat(service_dfs, ignore_index=True)

        dfs.append(taxis_data)

    # Concatenar los DataFrames de amarillo y verde
    taxis = pd.concat(dfs, ignore_index=True)

    # Aplicar la función para crear nuevas columnas
    taxis['pickup_borough'] = taxis['PULocationID'].apply(get_borough)
    taxis['dropoff_borough'] = taxis['DOLocationID'].apply(get_borough)

    # Eliminar filas donde pickup_borough o dropoff_borough sea igual a 'Unknown'
    taxis = taxis[(taxis['pickup_borough'] != 'Unknown') & (taxis['dropoff_borough'] != 'Unknown')]

    # Cambiar el tipo de dato de 'passenger_count' a int
    taxis['passenger_count'] = taxis['passenger_count'].astype(int)

    # Guardar el DataFrame en un archivo parquet dentro del bucket de S3
    taxis.to_parquet("s3://web-scrap-bucket/taxis_2023.parquet")
    print("ETL process completed.")
