==> 1-Local/6-AWS_Endpoints.ipynb <==
# **Funciones para Endpoints en AWS**

#### **Elegimos realizar nuestras operaciones de Big Data con [`AWS`](https://aws.amazon.com/es/) debido a que es la empresa líder en el sector y cuenta con una alta demanda en la industria. La decisión se basa en la reputación y experiencia sobresalientes de [`AWS`](https://aws.amazon.com/es/) en la gestión eficiente de datos a gran escala. Optar por [`AWS`](https://aws.amazon.com/es/) nos permite aprovechar su infraestructura sólida, servicios avanzados y la confianza que inspira en el mercado, asegurando así que nuestras operaciones de Big Data estén respaldadas por una plataforma de clase mundial reconocida por su fiabilidad y eficacia.**

### Importamos las librerías necesarias para el desarrollo de las funciones.

import boto3
import pandas as pd
import pyarrow.parquet as pq
import pyarrow.fs
from s3fs import S3FileSystem
import s3fs
import joblib
from fastapi import FastAPI, Depends
from pydantic import BaseModel
from datetime import timedelta
from pyathena import connect

### Probamos la conexión con el bucket.

# Especificamos el nombre de tu bucket y el identificador de la región de AWS
bucket_name = 'tu_bucket'
region_name = 'tu_region'
aws_access_key_id = 'tu_access_key_id'
aws_secret_access_key = 'tu_secret_access_key'

# Creamos un cliente de S3 con las credenciales proporcionadas
s3 = boto3.client('s3', region_name=region_name, aws_access_key_id=aws_access_key_id, aws_secret_access_key=aws_secret_access_key)

# Listamos objetos en el bucket
response = s3.list_objects(Bucket=bucket_name)
for obj in response.get('Contents', []):
    print(f'Archivo: {obj["Key"]}, Tamaño: {obj["Size"]} bytes')

### Revisamos el contenido de un archivo del bucket al azar.

# Especificamos las credenciales de AWS
aws_access_key_id = 'tu_access_key_id'
aws_secret_access_key = 'tu_secret_access_key'

# Especificamos la ruta de tu archivo Parquet en S3
s3_path = 'tu_bucket/tu_archivo'

# Configuramos la conexión a S3 con las credenciales
s3 = s3fs.S3FileSystem(key=aws_access_key_id, secret=aws_secret_access_key)

# Abrimos el archivo Parquet directamente desde S3
parquet_file = pq.ParquetFile(s3.open(s3_path))

# Leemos el archivo Parquet en un DataFrame de pandas
df = parquet_file.read().to_pandas()

# Imprimimos los resultados
print(df.info())

#### `Función que devuelve la predicción porcentual de oportunidad de viajes de un usuario.`

class InputData(BaseModel):
    pBoroughID: int
    dayofweek: int
    hour: int


def predict(data: InputData):

    # Especificamos las credenciales de AWS
    aws_access_key_id = 'tu_access_key_id'
    aws_secret_access_key = 'tu_secret_access_key'

    # Configuramos la conexión a S3 con las credenciales
    s3 = s3fs.S3FileSystem(key=aws_access_key_id, secret=aws_secret_access_key)

    # Especificamos la ruta del modelo en S3
    model_path = 'tu_bucket/tu_modelo'

    # Descargamos el modelo desde S3
    with s3.open(model_path, 'rb') as f:
        xgb_model = joblib.load(f)

    # Convertirmos los datos de entrada a un DataFrame
    input_data = pd.DataFrame([data.model_dump()])

    # Realizarmos la predicción
    prediction = xgb_model.predict(input_data)

    # Devolvemos la predicción como JSON
    return {"Probabilidad de conseguir pasajero": prediction.tolist()}

input_data = InputData(pBoroughID=1, dayofweek=5, hour=18)
predict(input_data)

#### `Función que devuelve un top 3 de vehículos ecológicos de acuerdo al capital ingresado`

def top_3_vehicles(max_price_usd: float):

    # Especificamos las credenciales de AWS
    aws_access_key_id = 'tu_access_key_id'
    aws_secret_access_key = 'tu_secret_access_key'

    # Especificamos la ruta del archivo Parquet en S3
    s3_path = 'tu_bucket/tu_archivo'

    # Configura la conexión a S3 con las credenciales
    s3 = s3fs.S3FileSystem(key=aws_access_key_id, secret=aws_secret_access_key)

    # Abrimos el archivo Parquet directamente desde S3
    parquet_file = pq.ParquetFile(s3.open(s3_path))

    # Leemos el archivo Parquet en un DataFrame de pandas
    df = parquet_file.read().to_pandas()

    # Filtramos los vehículos por precio
    filtered_vehicles = df[df['Price (USD)'] <= max_price_usd]

    # Ordenamos los vehículos por CO2 en orden ascendente, luego por Sound Emission y finalmente por Range
    sorted_vehicles = filtered_vehicles.sort_values(by=['CO2 Emission (g/mi)', 'Sound Emission (dB)', 'Range (mi)'], ascending=[True, True, False])

    # Tomamos los tres primeros vehículos dentro del rango de precio
    top_3_vehicles = sorted_vehicles.head(3)

    # Creamos el formato de salida como una lista de diccionarios
    output_format = []
    for idx, row in enumerate(top_3_vehicles.itertuples(), start=1):
        output_format.append({
            f'Puesto {idx}': f'{row.Manufacturer} {row.Model}',
            'Precio (USD)': row._9,
            'Combustible': row.Fuel,
            'CO2': row._7,
            'dB': row._8,
            'Millas con un tanque lleno': row._10,
        })

    return output_format

top_3_vehicles(25500)

#### `Función que devuelve estadísticas de viajes de acuerdo al Borough ingresado`

def taxi_stats_test(pickup_borough: str):
    # Configuramos las credenciales de AWS
    aws_access_key_id = 'tu_access_key_id'
    aws_secret_access_key = 'tu_secret_access_key'
    region_name = 'tu_region'

    # Configuramos la conexión a Athena
    conn = connect(aws_access_key_id=aws_access_key_id,
                aws_secret_access_key=aws_secret_access_key,
                region_name=region_name,
                schema_name='tu_schema',
                s3_staging_dir='tu_bucket')

    # Ejecutamos una consulta
    query = '''
        WITH tu_table AS (
    SELECT
        "pickup_borough",
        COUNT(*) AS "Viajes_Totales",
        ROUND(AVG(date_diff('second', "pickup_datetime", "dropoff_datetime")) / 3600, 2) AS "Duracion_Promedio_Hs",
        ROUND(COUNT(*) / COUNT(DISTINCT date_trunc('day', "pickup_datetime")), 2) AS "Media_de_Viajes_por_Dia",
        ROUND(COUNT(*) / COUNT(DISTINCT date_trunc('month', "pickup_datetime")), 2) AS "Media_de_Viajes_por_Mes",
        ROUND(AVG("trip_distance"), 2) AS "Distancia_recorrida_promedio_millas",
        ROUND(AVG("total_amount"), 2) AS "Total_ganado_en_promedio_USD"
    FROM
        "tu_database"."tu_table"
    GROUP BY
        "pickup_borough"
    )
    SELECT
    *
    FROM
    tu_table
    ORDER BY
    "pickup_borough"
    '''

    cursor = conn.cursor()
    cursor.execute(query, parameters={'pickup_borough': 'Manhattan'})  

    result = pd.DataFrame(cursor.fetchall(), columns=["pickup_borough", "Viajes_Totales", "Duracion_Promedio_Hs", "Media_de_Viajes_por_Dia", "Media_de_Viajes_por_Mes", "Distancia_recorrida_promedio_millas", "Total_ganado_en_promedio_USD"])

    borough_result = result[result["pickup_borough"] == pickup_borough]

    formatted_result = {
        "Viajes Totales": borough_result["Viajes_Totales"].values[0],
        "Duración Promedio (Hs)": borough_result["Duracion_Promedio_Hs"].values[0],
        "Media de Viajes por Día": borough_result["Media_de_Viajes_por_Dia"].values[0],
        "Media de Viajes por Mes": borough_result["Media_de_Viajes_por_Mes"].values[0],
        "Distancia recorrida promedio (millas)": borough_result["Distancia_recorrida_promedio_millas"].values[0],
        "Total ganado en promedio (USD)": borough_result["Total_ganado_en_promedio_USD"].values[0]
    }

    # En el entorno de Athena, enfrentamos un inconveniente al realizar consultas desde Python, ya que dichas consultas generan
    # automáticamente dos archivos en el bucket: uno con extensión .csv y otro con extensión .metadata. Esta situación impide la
    # ejecución de nuevas consultas hasta que ambos archivos sean eliminados. Para superar este problema, hemos implementado una
    # solución que conecta con S3 añadiendo las siguientes líneas de código. Estas líneas permiten la eliminación de los archivos generados,
    # posibilitando así la ejecución de consultas sin inconvenientes.
    
    s3_client = boto3.client('s3',
                            aws_access_key_id=aws_access_key_id,
                            aws_secret_access_key=aws_secret_access_key,
                            region_name=region_name)

    # Definimos el nombre del bucket
    bucket_name = 'tu_bucket'

    # Obtenemos la lista de objetos en el bucket
    response = s3_client.list_objects_v2(Bucket=bucket_name)

    # Eliminamos los archivos .csv y .metadata encontrados
    for obj in response['Contents']:
        if obj['Key'].endswith('.csv') or obj['Key'].endswith('.metadata'):
            s3_client.delete_object(Bucket=bucket_name, Key=obj['Key'])


    return formatted_result

taxi_stats_test('Manhattan')
